feat: Refactor RequestsView to PendingPanelistsView
- The grid now shows Panelist entities with 'pending' status.
- The 'new request' button and the edit form now work with the Panelist entity.
- Deleted the unused Request entity, repository, and service.
- Updated the main layout to navigate to the new view.
- Configured the application to use an in-memory H2 database for testing.
