--- src/main/java/uy/com/equipos/panelmanagement/data/RequestRepository.java
@@ -1,8 +1,0 @@
-package uy.com.equipos.panelmanagement.data;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-
-public interface RequestRepository extends JpaRepository<Request, Long>, JpaSpecificationExecutor<Request> {
-
-}
