+++ src/main/resources/application.properties
@@ -4,16 +4,17 @@ spring.mustache.check-template-location = false
 
 # Launch the default browser when starting the application in development mode
 vaadin.launch-browser=true
-# MariaDB configuration.
-spring.datasource.url = jdbc\:mariadb\://localhost\:3306/paneles
-spring.datasource.username = vaadinstart
-spring.datasource.password = vaadinstart
+# H2 configuration
+spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.ddl-auto = update
 # To improve the performance during development.
 # For more information https://vaadin.com/docs/latest/integrations/spring/configuration#special-configuration-parameters
 vaadin.allowed-packages = com.vaadin,org.vaadin,com.flowingcode,uy.com.equipos.panelmanagement
 spring.jpa.defer-datasource-initialization = true
 spring.sql.init.mode = always
-spring.jpa.database-platform=org.hibernate.dialect.MariaDBDialect
 alchemer.api.token=73305b19aec9ea61725b082179b9556ae32f494efc37e5b113
 alchemer.api.token.secret=A9lwGSTYjI0bQ
